requirements.txt
google_api_python_client==2.33.0
google_auth_oauthlib==0.4.2
httplib2==0.19.0
protobuf==3.19.1
requests==2.25.1